+++ changelog.md
@@ -1,8 +1,7 @@
 # Changelog 1.0.16
 
 TODO (Build):
-  - make a way to percentage + random on the blocks filled (maybe the wand make it have modes)
-  - make it cancelable, store the previous state of the blocks (with //undo + //redo) [BuildManager to complete]
+  - make noise values
 
 TODO (WorldGen):
   - priority: add animals & mobs inside the `dense_forest` biome (bears - detail: have something in their mouth sometimes)
@@ -145,7 +144,7 @@ New:
   - feat: added `dense_forest` biome
   - feat: added `bear` spawn with variants (tameable with various `sitting_poses`)
   - feat: jobs `tasks_progress` are now shared between players for competition
-  - feat: `/wand`, `//fill <block>` commands to help build structures (creative only) + visuals
+  - feat: `//wand`, `//fill <block> <meta> [parameters]`, `//undo`, `//redo`, `//randomize <block>:<meta>:<weight> ...` commands to help build structures (creative only) + visuals
 
 Changes:
   - feat: `oxonium_pickaxe` has `vein miner` mode compatible with other mods `ores`